+++ maker/src/types/text_renderer.rs
@@ -41,7 +41,7 @@ impl Renderer for TextRenderer {
         let columns = (self.text.len() as f32).sqrt().ceil() as usize;
         let rows = (self.text.len() as f32 / columns as f32).ceil() as usize;
         let data = alphabet::default::ALPHABET
-            .render(&self.text, columns, 16 * maze.width())
+            .render(&self.text, columns, 16 * maze.width() as usize)
             .map(|(pos, v)| {
                 (
                     physical::Pos {
