+++ maze/benches/walk.rs
@@ -8,8 +8,8 @@ pub fn walk(c: &mut Criterion) {
         for shape in [Shape::Tri, Shape::Quad, Shape::Hex].iter() {
             let maze =
                 Maze::<()>::new(black_box(*shape), 100, 100).initialize(method, &mut LFSR::new(65));
-            let start = (0isize, 0isize).into();
-            let end = ((maze.width() - 1) as isize, (maze.height() - 1) as isize).into();
+            let start = (0, 0).into();
+            let end = ((maze.width() - 1) as i32, (maze.height() - 1) as i32).into();
             group.bench_with_input(BenchmarkId::from_parameter(shape), shape, |b, _| {
                 b.iter(|| {
                     maze.walk(start, end);
