+++ web/src/types/mod.rs
@@ -12,7 +12,7 @@ mod seed;
 pub use self::seed::*;
 
 /// The maximum nmber of rooms.
-const MAX_ROOMS: usize = 1000;
+const MAX_ROOMS: u32 = 1000;
 
 /// A responder providing an image of a maze.
 pub struct Maze {
@@ -46,8 +46,8 @@ impl From<Maze> for HttpResponse {
                         maze.walk(
                             maze::matrix::Pos { col: 0, row: 0 },
                             maze::matrix::Pos {
-                                col: maze.width() as isize - 1,
-                                row: maze.height() as isize - 1,
+                                col: maze.width() as i32 - 1,
+                                row: maze.height() as i32 - 1,
                             },
                         )
                         .unwrap()
